pyproject.toml
[tool.poetry]
name = "boostmot-backend"
version = "0.1.0"
description = ""
authors = ["Uki Hong <suhong9507@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "boostmot_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
beautifulsoup4 = "^4.11.2"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-environ = "^0.10.0"
django-cors-headers = "^3.14.0"
requests = "^2.29.0"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
